requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.0.7
djangorestframework==3.12.2
gunicorn==20.0.4
mock==4.0.3
parameterized==0.8.1
psycopg2-binary==2.8.5
python-decouple==3.4
pytz==2019.3
six==1.12.0
sqlparse==0.3.1
whitenoise==5.0.1
